+++ jwt-spring/src/main/java/jwtspring/service/userImg/UserImgService.java
@@ -1,9 +1,8 @@
 package jwtspring.service.userImg;
 
-import jwtspring.models.UserProfileImage;
+import jwtspring.models.user.UserProfileImage;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -12,7 +11,7 @@ import java.io.IOException;
 @Service
 public interface UserImgService {
 
-    ResponseEntity uplaodImage(@RequestParam("imageFile") MultipartFile file, @RequestParam("userId") long userid) throws IOException;
+    ResponseEntity uplaodImage(@RequestParam("imageFile") MultipartFile file, @RequestParam("userId") int userid) throws IOException;
 
     UserProfileImage getImage(int userImgId) throws IOException;
 }
